+++ src/Runner/Runner.php
@@ -70,7 +70,7 @@ class Runner
     )
     {
         $flags = null === $flags ? new Flags() : $flags;
-        $env = null === $env ? Env::create() : $env;
+        $env = null === $env ? Env::createEx() : $env;
         $streams = null === $streams ? Streams::create() : $streams;
 
         return new self($runOpts, $directories, $exec, $flags, $env, $streams);