+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 ### Fixed
+- Help message on trigger destination branch name
 - Remove static calls to throw a status exception
 
 ## [0.0.35] - 2020-05-24